+++ mail_2_rmq.py
@@ -186,6 +186,35 @@ def archive_email(rq, log, cfg, msg, **kwargs):
     log.log_info("Email saved to:  %s" % (e_file))
 
 
+def get_text(msg, **kwargs):
+
+    """Function:  get_text
+
+    Description:  Walks the tree of a email and returns the text of the email.
+
+    Arguments:
+        (input) msg -> Email message instance.
+        (input) **kwargs:
+            None
+        (output) All texts in email joined together in a single string.
+
+    """
+
+    msg_list = []
+
+    for part in msg.walk():
+
+        if part.get_content_maintype() == "multipart":
+            continue
+
+        elif not part.get_payload(decode=True):
+            continue
+
+        msg_list.append(part.get_payload(decode=True))
+
+    return "".join(msg_list)
+
+
 def connect_process(rq, log, cfg, msg, **kwargs):
 
     """Function:  connect_process