+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [0.3.4] - 2019-02-19
+### Added:
+- get_text:  Walks the tree of a email and returns the text of the email.
+
 ### Changed
 - parse_email:  Replaced "email" class with "email.Parser" class.